extract crash processing code
